backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
openai
httpx<0.28
numpy
motor
python-jose[cryptography]
beautifulsoup4
mammoth
python-docx
pdf2image
pillow
pymupdf
transformers
transformers
torch
